pyproject.toml
[tool.poetry]
name = "cli-surf"
version = "0.1.0"
description = ""
authors = ["ryansurf <your@email.com>"] # TODO: email
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
# command: `poetry add <package-name>`
python = ">3.9.7"
flask = "3.0.3"
flask-cors = "5.0.0"
g4f = "0.3.2.2"
geopy = "2.4.1"
openmeteo-requests = "1.2.0"
pandas = "2.2.2"
pydantic = "2.7.2"
pydantic-settings = "2.2.1"
python-dotenv = "1.0.1"
requests = "2.32.4"
requests-cache = "1.2.0"
retry-requests = "2.0.0"
email-validator = "2.1.1"
openai = "^1.30.5"
curl-cffi = "^0.7.1"
streamlit = "^1.35.0"
streamlit-folium = "^0.22.0"
matplotlib = "3.9.0"
seaborn = "^0.13.2"
numpy = "^1.26.4"
pymongo = {extras = ["gssapi", "snappy", "srv", "tls"], version = "^4.15.4"}

[tool.poetry.group.dev.dependencies]
# command: `poetry add --group dev <package-name>`
ruff = "0.4.6"
pytest = "8.2.1"
pytest-mock = "*"
pytest-cov = "5.0.0"
pre-commit = "3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# pytest settings
[tool.pytest.ini_options]
testpaths = 'tests'
pythonpath = "."
addopts = '-p no:warnings'  # disable pytest warnings
log_format = '%(name)s %(levelname)s: %(message)s'

# coverage settings
[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
omit = ["src/dev_streamlit.py"]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]

# ruff global settings
[tool.ruff]
line-length = 79

# linter
[tool.ruff.lint]
# I: isort
# F: Pyflakes
# E: Pycodestyle Error
# W: Pycodestyle Warning
# P: Pylint
# PT: flake8-pytest-style
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

# formatter
[tool.ruff.format]
preview = true
quote-style = 'double'
